% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatParameter.R
\name{CatParameter}
\alias{CatParameter}
\alias{filter_occupancy}
\title{CatParameter S3 class}
\usage{
CatParameter(x)

filter_occupancy(x, i)
}
\arguments{
\item{x}{A named list of \code{data.frames}, each being one parameter table (see Details for more information)}
}
\description{
Create S3 class for CatParameter data type
}
\details{
CatParameter usually include parameters as follows:
\itemize{
  \item Occupancy
  \item Coverage
  \item ConstructionType
  \item BuildingHeight
  \item YearBuilt
  \item LossTerm (e.g. Deductible and Sublimit)
}

Above are also valid names for input \code{x}
}
\section{Functions}{
\itemize{
\item \code{filter_occupancy}: Get a subset of \code{CatParameter} with specified Occupancy types
}}

\examples{
prm1 <- CatParameter(param_long)
prm1
prm2 <- CatParameter(param_wide)
all.equal(prm1, prm2)
filter_Occupancy(prm1, c("Commercial", "Industrial"))

}
