% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTester.R
\name{cli_check_that}
\alias{cli_check_that}
\title{In-line checker}
\usage{
cli_check_that(code, desc = NULL)
}
\description{
In-line checker
}
\examples{
cli_check_that(nrow(mtcars) == 32)
cli_check_that(nrow(mtcars) == 31, "# of rows for `mtcars` is 31")

}
