% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnames.R
\name{cnames}
\alias{cnames}
\title{Helper function to get matched cols of a data.table}
\usage{
cnames(DT, cols = NULL)
}
\arguments{
\item{DT}{A \code{data.table} or \code{data.frame} object}

\item{cols}{Fully compatible with \code{data.table}'s \code{.SDcols} parameter}
}
\description{
Helper function to get matched cols of a data.table
}
\examples{
cnames(mtcars) # same as names(mtcars)
cnames(mtcars, 1:3) # same as names(mtcars)[1:3]
cnames(mtcars, !c("mpg", "cyl"))
cnames(mtcars, is.character)
cnames(mtcars, !is.character)
cnames(mtcars, vs:carb)
cnames(mtcars, !patterns("mpg|cyl")) # regular expression
cnames(mtcars, list(cyl:hp, patterns("carb"), !is.integer)) # union of results for list input
cnames(mtcars, .(cyl:hp, patterns("carb"), !is.integer)) # `.` is also interpreted as list
ctypes(mtcars)
ctypes(mtcars, as.DT = TRUE)
# use cnames in a function
cnames2 <- function(DT, col = NULL) {
  cc <- substitute(col)
  cnames(DT, !!cc)
}
cnames2(mtcars)
cnames2(mtcars, vs:carb)
cnames2(mtcars, !patterns("mpg|cyl")) # regular expression
}
