% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{color_contrast}
\alias{color_contrast}
\title{Decide whether to use light color or dark color according to color palette in \code{x}}
\usage{
color_contrast(x, col_light = "#ffffff", col_dark = "#000000")
}
\description{
Usually \code{x} is the background color in plot, and the output is the font color
}
