% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-Contract.R
\name{ct_Excess}
\alias{ct_Excess}
\alias{ct_LossParticipation}
\alias{ct_ProfitCommission}
\alias{ct_Commission}
\title{Common contract terms}
\usage{
ct_Excess(
  Loss,
  Premium = NULL,
  Limit,
  Deductible,
  is.FranchiseLimit = FALSE,
  is.FranchiseDed = FALSE
)

ct_LossParticipation(Loss, Premium, Term, is.CedentPerspective = TRUE)

ct_ProfitCommission(LossAndExpense, Premium, Term, ExpenseAllowance)

ct_Commission(Loss = NULL, Premium, Term, right.closed = TRUE)
}
\arguments{
\item{Loss, Premium}{Numeric vector for loss and premium}

\item{Limit, Deductible}{Limit and Deductible terms}

\item{is.FranchiseLimit}{Logical, franchise limit means there is no reimbursement when LossAboveDeducible is greater than Limit}

\item{is.FranchiseDed}{Logical, franchise deductible means when Loss is greater than Deductible, full f.g.u. loss is applicable for reimbursement}

\item{Term}{See Details}

\item{is.CedentPerspective}{Logical.  If \code{TRUE}, the output is negative representing the loss amount that the cedent will bear}

\item{right.closed}{Logical, default to be TRUE.  RightBound inclusive}
}
\description{
\code{ct_Excess} calculates LossInLayer based on specified Limit and Deductible structure
\code{ct_Commission} calculates commission figures based on given sliding scale table
\code{ct_Reinstatement} returns a list of two vectors, one being reinstated loss and the other is reinstated premium
\code{ct_ProfitCommission} calculates profit commission
\code{ct_LossParcipation} calculates loss participation (i.e. loss corridor) from cedent's perspective
}
\section{\code{Term} parameter for different functions}{

\itemize{
\item Commission:  parameter \code{Term} can either be a value (for fixed commission), or a 3-column \code{data.frame}, with the sequence of \code{LowerLR}, \code{UpperLR} and \code{Commission}
\item LossParticipation: parameter \code{Term} is 3-column \code{data.frame}, with the sequence of \code{LowerLR}, \code{UpperLR} and \code{Participation}
\item ProfitCommission: parameter \code{Term} either a number a 3-column data.frame, each representing \code{LowerProfit}, \code{UpperProfit} and \code{ProfitShare}.
}
}

\examples{
ct_ProfitCommission(c(.2, .6, .8), ProfitShare = .2, ExpenseAllowance = .2)

SS <- data.table::data.table(LowerLR = c(0, .15, .2, .25), UpperLR = c(.15, .2, .25, Inf), Commission = c(.3, .25, .2, .15))
ct_Commission(Loss = c(.19, .2, .21), Premium = 1, SS)
ct_Commission(Loss = c(.19, .2, .21), Premium = 1, Term = .33)
}
