% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTester.R
\name{DataTester}
\alias{DataTester}
\title{A container for all data tests}
\description{
A container for all data tests

A container for all data tests
}
\examples{
dt1 <- DataTester$new()
DT <- data.table::as.data.table(mtcars)
dt1$test_that("cyl is below 20", {DT[cyl > 20]})
dt1$test_that("cyl cannot be 6", {DT[cyl == 6]}, "cyl2")
dt1$test_that("There are 32 rows", {nrow(DT) == 32})
## or wrap them all into a function
test_f <- function() {
  dt1 <- DataTester$new()
  dt1$test_that("cyl is below 20", {DT[cyl > 20]})
  dt1$test_that("cyl cannot be 6", {DT[cyl == 6]}, "cyl2")
}
test_f()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataTester$new()}}
\item \href{#method-test_that}{\code{DataTester$test_that()}}
\item \href{#method-summary}{\code{DataTester$summary()}}
\item \href{#method-write_to_excel}{\code{DataTester$write_to_excel()}}
\item \href{#method-print}{\code{DataTester$print()}}
\item \href{#method-keep_cols}{\code{DataTester$keep_cols()}}
\item \href{#method-clone}{\code{DataTester$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new DataTester object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTester$new(cols_keep = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols_keep}}{A character vector for columns to keep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_that"></a>}}
\if{latex}{\out{\hypertarget{method-test_that}{}}}
\subsection{Method \code{test_that()}}{
Create a new test
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTester$test_that(desc, code, test_name = NULL, cols_keep = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{desc}}{Human-readble description for the test}

\item{\code{code}}{Code for testing, can be embraced with curly bracket (i.e. {...})}

\item{\code{test_name}}{Concise test id, usually for outputing as Excel sheetname.  If left blank, an id like "test-001" will be created automatically}

\item{\code{cols_keep}}{Columns in output dataset (for invalid records).  This will over-rule the cols_keep input during object initialization}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Summarise the results for all tests
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTester$summary(invalid_only = TRUE, sorted = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{invalid_only}}{Logical, only show invalid tests}

\item{\code{sorted}}{Logical, whether to reorder the tests by \code{test_name}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_to_excel"></a>}}
\if{latex}{\out{\hypertarget{method-write_to_excel}{}}}
\subsection{Method \code{write_to_excel()}}{
Output test summary and invalid datasets to Excel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTester$write_to_excel(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{File path for excel output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
For S3 print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTester$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keep_cols"></a>}}
\if{latex}{\out{\hypertarget{method-keep_cols}{}}}
\subsection{Method \code{keep_cols()}}{
Overwrite keep_cols on the flight
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTester$keep_cols(cols_keep = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols_keep}}{A character vector for columns to keep}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTester$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
