% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{date_duration}
\alias{date_duration}
\title{Calculate duration between two dates}
\usage{
date_duration(d1, d2, unit = c("day", "month", "quarter", "year"))
}
\arguments{
\item{d1, d2}{vector of \code{Date} object}

\item{unit}{One of month, quarter or year}
}
\description{
Use month as basis.  Treat one "standard" month as of 30 days. Number of months are rounded to integer. e.g. 29 days are treated as 1 month, 1/3 quarter and 1/12 year
}
\examples{
d <- as.Date(c("2010-03-31", "2010-06-30", "2011-09-30"))
date_duration(as.Date("2010-01-01"), d, "month")
date_duration(as.Date("2010-01-01"), d, "quarter")
date_duration(as.Date("2010-01-01"), d, "year")
}
