% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{date_parse_excel}
\alias{date_parse_excel}
\alias{date_parse_month}
\alias{date_parse_quarter}
\title{Convert excel date integer into R date}
\usage{
date_parse_excel(x, ...)

date_parse_month(x, period_end = FALSE)

date_parse_quarter(x, period_end = TRUE)
}
\description{
Convert excel date integer into R date
}
\examples{
date_parse_excel(c(NA, 43831, 43466))
x_month <- c(NA, "Aug.2005", "Jan to Feb 2008", "Mid September, 2018", "2018 July", "2019/07")
date_parse_month(x_month)
date_parse_month(x_month, period_end = TRUE)
x_quarter <- c(NA, "2019/Q3", "2018Q3", "Q2-2016", "1Q/15")
date_parse_quarter(x_quarter)
date_parse_quarter(x_quarter, period_end = FALSE)
}
