% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriDev.R
\name{dev_CapeCod}
\alias{dev_CapeCod}
\title{Complete the input triangle using CapeCod B-F method}
\usage{
dev_CapeCod(tri_loss, ultimate_premium, ELR = NULL, ...)
}
\description{
Complete the input triangle using CapeCod B-F method
}
\examples{
prem <- as_triangle(TRI$Premium)
ultprem <- dev_ChainLadder(prem)[, ncol(prem) + 1]
loss <- as_triangle(TRI$Incurred)
dev_CapeCod(loss, ultprem)
}
