% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{doughnut}
\alias{doughnut}
\title{Plot doughnut chart in base R (with specs mirroring the \code{pie} function)}
\usage{
doughnut(
  x,
  labels = names(x),
  edges = 200,
  outer.radius = 0.8,
  inner.radius = 0.5,
  clockwise = FALSE,
  init.angle = if (clockwise) 90 else 0,
  density = NULL,
  angle = 45,
  col = NULL,
  border = FALSE,
  lty = NULL,
  main = NULL,
  ...
)
}
\description{
Slightly modified by based on codes in \link{https://r-graph-gallery.com/130-ring-or-donut-chart.html}
}
\examples{
doughnut(c(3, 5, 9, 12), labels = LETTERS[1:4], col = palette_op("economist")[c(1, 3, 5, 7)])
}
