% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_extra.R
\name{dt_nest}
\alias{dt_nest}
\title{Nest a data.table}
\usage{
dt_nest(DT, by, key_name = "data", keep = FALSE)
}
\arguments{
\item{DT}{A \code{data.table}}

\item{by}{Var names used for grouping}

\item{key_name}{Name for the list-column}

\item{keep}{Logical, whether to keep \code{by} columns in the list-column}
}
\description{
Nest a data.table
}
\examples{
x <- data.table::data.table(Id  = c("A", "A", "C", "C"),
                X1  = c(1L, 3L, 5L, 7L),
                XY  = c("x2", "x4", "x6", "x8"),
                key = "Id")
x |> dt_nest(by = "Id")
x |> dt_nest(by = "Id", keep = TRUE)
}
