% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_extra.R
\name{dt_rep}
\alias{dt_rep}
\title{Replicate a \code{data.table} (vertically)}
\usage{
dt_rep(DT, times = NULL, names = NULL, idcol = ".ID")
}
\description{
Replicate a \code{data.table} (vertically)
}
\examples{
DT <- data.table::data.table(x = LETTERS[1:3], y = 5:7)
dt_rep(DT, 3)
dt_rep(DT, names = c("2021", "2022", "2023"), idcol = "Year")
}
