% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_extra.R
\name{dt_set}
\alias{dt_set}
\title{Lazy-evaluation version of `:=`}
\usage{
dt_set(DT, ...)
}
\description{
Variable in DT is prefered than that in parent.frame
}
\examples{
DT <- as.data.table(mtcars)
A <- 2
X <- 3
dt_set(copy(DT), A = mpg, B = A * 2, C = X / 3)
dt_set(copy(DT), A = mpg, B = mean(max(A * 2)), A = NULL, mpg = NULL)
}
