% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_extra.R
\name{dt_setcols}
\alias{dt_setcols}
\title{Helper function to apply a function to multiple cols}
\usage{
dt_setcols(DT, cols, FUN, ...)
}
\arguments{
\item{DT}{A \code{data.table} or \code{data.frame} object}

\item{cols}{Fully compatible with \code{data.table}'s \code{.SDcols} parameter}
}
\description{
Helper function to apply a function to multiple cols
}
\examples{
x <- data.table::data.table(A = LETTERS[1:3], B = 5:7, C = 9:11)
dt_setcols(copy(x), c("B", "C"), function(x) x * 100)[]
dt_setcols(copy(x), 2:3, function(x) x * 100)[]
dt_setcols(copy(x), B:C, function(x) x * 100)[]
dt_setcols(copy(x), is.numeric, function(x) x * 100)[]
dt_setcols(copy(x), patterns("B|C"), function(x) x * 100)[]
dt_setcols(copy(x), !is.character, function(x) x * 100)[]
dt_setcols(copy(x), .(2:3, "B"), function(x) x * 100)[]
}
