% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{elapply}
\alias{elapply}
\title{Feed elements in a list to a function}
\usage{
elapply(X, FUN, ...)
}
\arguments{
\item{X}{A list}

\item{FUN}{A function}

\item{...}{optional arguments to \code{FUN}}
}
\description{
Feed elements in a list to a function
}
\examples{
x <- list(a = 1:3, b = 4:6)
f <- function(a, b, mult = 1) (a + b) * mult
elapply(x, f) ## 5, 7, 9
elapply(x, f, mult = 2) ## 10, 14, 18
elapply(x[1], f, b = 1) ## 2, 3, 4
}
