% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-CurveFitting.R
\name{fit_severity}
\alias{fit_severity}
\title{Fit conditional severity distribution.}
\usage{
fit_severity(loss, distribution, Min, Max, wt = NULL)
}
\arguments{
\item{loss}{A numeric vector}

\item{wt}{(Optional) a numeric vector with same length as \code{loss}}
}
\value{
A named vector for parameters
}
\description{
Fit conditional severity distribution.
}
\examples{
\dontrun{
a <- c(2128626.826, 1001231.566, 2602453.407, 2652625.402, 2096552.901, 2481345.789, 1767929.774, 1132111.303, 1562121.048, 1529111.294, 1875538.967, 3671208.834, 1566929.993, 1900074.241, 1170474.947, 1958780.493, 1226562.609, 2364993.783, 4393903.605, 1711382.839, 3246468.782, 2749344.86, 2325670.505, 1094377.265, 1124804.947, 1098723.109, 1835881.542, 1602722.045, 1569878.823, 1024398.477, 2156012.186, 2644855.515, 1760780.099, 3764595.472, 1429465.965, 2008155.44, 1967520.282, 1954653.638, 6524999.764, 1839039.868, 2292137.751, 1631507.991, 1063858.728, 1061057.367, 1624566.74, 1200184.015, 2565640.453, 1230331.532, 1523257.634, 3314405.795, 2171018.269, 1551589.011, 1696161.672, 2262381.227)

fit_severity(c(100, 200, 150, 300), Min = 100, Max = 500, distribution = "SimplePareto")
}
}
