% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPT_import_PRTemplate.R
\name{import_PRTemplate_Dev}
\alias{import_PRTemplate_Dev}
\title{Read from GC Pro-Rata Template to get Dev table}
\usage{
import_PRTemplate_Dev(
  wb,
  LOB,
  LRSummary,
  Projection = LOB,
  Large = "Large",
  Event = "CAT"
)
}
\arguments{
\item{wb}{Path of GC PR Treaty template workbook}

\item{LOB}{LOB name defined in template}

\item{LRSummary}{Name for LossRatioSummary sheet}

\item{Projection}{Name for Projection sheet}

\item{Large}{Name for LargeLossList sheet}

\item{Event}{Name for EventLossList sheet}
}
\description{
Read from GC Pro-Rata Template to get Dev table
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
import_PRTemplate_Dev(wb, "Property", "Loss Ratios")
import_PRTemplate_Dev(wb, c("Property", "MB"), "Loss Ratios")
\dontshow{\}) # examplesIf}
}
