% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{kable_op}
\alias{kable_op}
\title{Basic format for kable}
\usage{
kable_op(
  DT,
  theme = c("green.fill", "blue.fill"),
  font_size = 14,
  header_sep = NULL,
  na_replace = "",
  cols_right = NULL
)
}
\description{
Basic format for kable
}
\examples{
x <- data.table::data.table(
  " " = "2020 Benchmark Price",
  "Terms_Limit" = c(10e6, 10e6, 20e6),
  "Terms_Deductible" = c(10e6, 20e6, 30e6),
  "Terms_Reinstatement" = c("2@100\%", "2@100\%", "1@100\%"),
  "Price_ROL" = c(.4, .35, .25)
  )
  kable_op(x) |> print()
  kable_op(x, header_sep = "_") |> print()
  kable_op(x, header_sep = "_", auto_num_format = TRUE) |> print()
  kable_op(x, split_header = TRUE) |> kableExtra::collapse_rows(columns = 1) |> print()
}
