% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{kable_save}
\alias{kable_save}
\title{Save formatted kable into png}
\usage{
kable_save(
  x,
  file,
  zoom = 2,
  bs_theme = "simplex",
  extra_dependencies = NULL,
  ...
)
}
\description{
Save formatted kable into png
}
