% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_format.R
\name{num}
\alias{num}
\alias{num_format}
\title{Format numbers}
\usage{
num(
  x,
  scale = c("auto", "free", "C", "P", "K", "M", "B", "pp"),
  digits = 2,
  na.as = NA_character_,
  inf.as = "Inf",
  ...
)

num_format(scale = c("auto", "free", "C", "P", "K", "M", "B"), digits = 0, ...)
}
\description{
Format numbers
}
\examples{
x <- c(1.23e12, 3.45e7, -6.78e3, Inf, NA)
num(x)
num(x, "free")
num(x, "free", inf.as = "Unlimited", na.as = "")
}
