% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical.R
\name{num_is_integer}
\alias{num_is_integer}
\title{Check if the elements of a numeric vector are close to integers}
\usage{
num_is_integer(x, tol = 1e-05)
}
\arguments{
\item{x}{Numeric vectors to check}

\item{tol}{Tolerance of comparison.}
}
\description{
Check if the elements of a numeric vector are close to integers
}
\examples{
num_is_integer(c(1.0000001, -1.9999999999, 5.99999999, -2, 2.123456, 2.0001, pi))
}
