% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onepunch-package.R
\docType{package}
\name{onepunch-package}
\alias{onepunch}
\alias{onepunch-package}
\title{onepunch: Personal toolkit for `data.table` based data wrangling and reinsurance analytical work}
\description{
A collections of my frequently-used functions and classes.
}
\author{
\strong{Maintainer}: Charles Pan \email{addnox@gmail.com}

}
\keyword{internal}
