% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPT_plot_PRTemplate.R
\name{op_plot_Dev}
\alias{op_plot_Dev}
\title{Plot Dev components}
\usage{
op_plot_Dev(
  x,
  y = NULL,
  var = NULL,
  palette = NULL,
  div_by = "UltPrem",
  num_scale = NULL,
  label_projected = NULL,
  label_format = NULL,
  geom_text_param = list(),
  scale_y_param = list(),
  var_plot_type = c("stack", "facet", "facet.with.total"),
  xy_label = c("Prior", "Current"),
  ...
)
}
\description{
Plot Dev components
}
\examples{
DT1 <- import_PRTemplate_Dev(wb = r"(C:\GC Projects\2024\PICC\NM SP\20231102_TreatyAnalysis_2023Q3_FinalData\PICC_2024_NMSP_Trending_20231105.xlsm)", LOB = c("Property", "MB", "CEAR"), LRSummary ="Smry_BeforeDed")[LOB == "Property"]
op_plot_Dev(DT1)
op_plot_Dev(DT1, label_format = num_format("P"), label_projected = "Proj.")
op_plot_Dev(DT1, label_format = num_format("P"), geom_text_param = list(nudge_y = .1, color = "#5bbcd6"))
op_plot_Dev(DT1, var = "ALC.Actual", label_projected = "Proj.")
op_plot_Dev(DT1, var = "ALC.Actual", var_plot_type = "facet.with.total", label_format = num_format("P", 1))
op_plot_Dev(DT1, var = "POI")
op_plot_Dev(DT1, var = "POI", scale_y_param = list(limits = c(0, 1.2), breaks = scales::breaks_width(.2)))
op_plot_Dev(DT1, DT1, var = "POI", xy_label = c("Q3 Study", "Q2 Study"), x_text_angle = 45)
}
