% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPT_plot_PRTemplate.R
\name{op_plot_Term}
\alias{op_plot_Term}
\title{Use \code{gt} package to lay out terms and structures}
\usage{
op_plot_Term(data, theme = "blue.fill")
}
\arguments{
\item{data}{A list of terms produced by \code{import_PRTemplate_Term}}
}
\description{
Use \code{gt} package to lay out terms and structures
}
