% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-GenericFunctions.R
\name{opMean}
\alias{opMean}
\alias{opSD}
\alias{opVaR}
\alias{opTVaR}
\title{Common summary statistics in reinsurance practice}
\usage{
opMean(x, ...)

opSD(x, ...)

opVaR(x, ...)

opTVaR(x, ...)
}
\description{
\code{opMean} mean
\code{opSD} standard deviation
\code{opVaR} Value at Risk
\code{opTVaR} Tail Value at Risk (a.k.a Conditional Tail Expectation, or CTE)
}
