% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_extra.R
\name{palette_op}
\alias{palette_op}
\title{A selection of color palettes}
\usage{
palette_op(name = c("gc", "economist"), idx = NULL)
}
\description{
A selection of color palettes
}
\examples{
palette_op() |> plot()
palette_op("economist", c(1, 3, 5, 6)) |> plot()
}
