% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{plot_save_base}
\alias{plot_save_base}
\title{Save base plot as png files}
\usage{
plot_save_base(plt, file, width = 9, height = 6)
}
\arguments{
\item{plt}{Plot object}

\item{file}{File name (with path)}

\item{width, height}{Width and height in inches}
}
\description{
Save base plot as png files
}
