% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_extra.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Plot waterfall chart}
\usage{
plot_waterfall(
  DT,
  balance_start = NULL,
  calc_balance = NULL,
  label_final = "Final Balance",
  color = c("#9ae5de", "#efe8d1", "#acc8d4"),
  color_border = "#7f7f7f",
  x_var = NULL,
  y_var = NULL,
  bar_width = 0.5,
  show_hline = FALSE,
  show_value = FALSE,
  labeller = NULL,
  ...
)
}
\arguments{
\item{DT}{A \code{data.frame} with at least two columns, usually the 1st for labels and the 2nd for incremental values}

\item{balance_start}{A named character vector with length 1.  The \code{value} will be the initial balance, and the \code{name} will be the label}

\item{calc_balance}{A named character vector.  The \code{value}s will be the position (after which) to calculate on-going balance.  The \code{name}s will be the labels}

\item{label_final}{The label for final balance}

\item{color}{Colors to fill the rectangles, with the sequence for "increase", "decrease" and "balance"}

\item{color_border}{Color for the rectangle border.  Use \code{NA} if no color is wanted}

\item{bar_width}{Width of rectangles}

\item{show_hline}{Show the link lines from the end of previous rectangle to the start of the next}

\item{show_value}{Show values as rectangle labels}

\item{labeller}{A function to format the value labels}

\item{...}{To be passed to \code{theme_op}}

\item{xvar, yvar}{Column names for \code{x} (i.e. label) and \code{y} (i.e. value).  If \code{NULL}, then the 1st column of \code{DT} will be used as label (if it is of character type), and the 2nd as value (if it is of numeric type)}
}
\description{
Plot waterfall chart
}
\examples{
df <- data.table(x = c("LR", "LPC", "Comm", "PC"), y = -c(.7, -.1, .5, 0))
plot_waterfall(df, show_hline = T, show_value = T)
plot_waterfall(df, balance_start = c("Premium" = 1), show_hline = T, show_value = T, labeller = num_auto)
plot_waterfall(df, balance_start = c("Premium" = 1), calc_balance = c("Bal after LPC" = "LPC", "Treaty Balance" = "PC"), show_hline = T, color = c("green", "red", "grey"), x_text_angle = 45, show_axis_text = "x")
}
