% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-ExposureRating.R
\name{PropExposureRating}
\alias{PropExposureRating}
\title{Create \code{S3} class for Property Exposure Rating (i.e. First Loss Scale method)}
\usage{
PropExposureRating(
  RiskProfile,
  SubjectPremium,
  ExpectLossRatio,
  RiskCurve,
  CurveType = NULL
)
}
\arguments{
\item{RiskProfile}{A \code{data.frame} containing at least the following 3 columns: \code{N}, \code{SI} and \code{Premium}}

\item{SubjectPremium}{Numeric, Subject premium to scale \code{RiskProfile}'s total premium}

\item{ExpectLossRatio}{Numeric, loss ratio to convert \code{SubjectPremium} to \code{ExpectLoss}}

\item{RiskCurve}{Created by \code{RiskCurve} function, or a risk curve function}
}
\description{
Create \code{S3} class for Property Exposure Rating (i.e. First Loss Scale method)
}
\details{
The \code{PropExposureRating} class support the following methods:
\itemize{
\item \code{loss_in_layer}
\item \code{build_cdf} (to do)
}
}
