% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxl.R
\name{readxl}
\alias{readxl}
\title{DT version of read_excel}
\usage{
readxl(
  wb,
  ws = NULL,
  rows = NULL,
  cols = NULL,
  range = NULL,
  col_names = TRUE,
  col_types = NULL,
  ...
)
}
\description{
DT version of read_excel
}
