% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxl.R
\name{readxl_raw}
\alias{readxl_raw}
\title{Read entire Excel sheet from cell \code{A1}}
\usage{
readxl_raw(
  wb,
  ws = NULL,
  rows = NULL,
  cols = NULL,
  range = NULL,
  col_names = FALSE,
  col_types = "text",
  alpha_type_cols = FALSE,
  ...
)
}
\description{
Read entire Excel sheet from cell \code{A1}
}
