% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{slapply}
\alias{slapply}
\title{Loop for side-effect only}
\usage{
slapply(X, FUN, ...)
}
\arguments{
\item{X}{Vector to loop through}

\item{FUN}{Function that is applied to \code{X}}

\item{...}{optional arguments to \code{FUN}}
}
\description{
Loop for side-effect only
}
\examples{
a <- slapply(1:3, function(x) cat("Num", x, "\n"))
a ## Return 1, 2 and 3, rather than Num 1, Num 2 and Num 3
}
