% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_split.R
\name{tidy_hsplit}
\alias{tidy_hsplit}
\alias{tidy_vsplit}
\title{Split data into list of \code{data.table}s horizontally or vertically}
\usage{
tidy_hsplit(x, by = NULL, na.rm = FALSE, keep.by = TRUE, factor.by = FALSE)

tidy_vsplit(x, by = NULL, na.rm = TRUE, keep.by = TRUE, factor.by = FALSE)
}
\description{
Split data into list of \code{data.table}s horizontally or vertically
}
