% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_split.R
\name{tidy_hsplit_}
\alias{tidy_hsplit_}
\title{Split data.table into list}
\usage{
tidy_hsplit_(x, cols = NULL, trim = FALSE, na.rm = TRUE, keep_cols = TRUE)
}
\arguments{
\item{x}{A data.frame}

\item{cols}{Numeric vector (indicating position) or character vector (indicating colnames)}

\item{trim}{Use \code{tidy_trim} for each element of output list}

\item{na.rm}{When TRUE, those rows with all \code{cols} being \code{NA} will be deleted before splitting}
}
\description{
tidy split functions return a list of split data blocks
}
\details{
\describe{
\item{position}{
\itemize{
\item \code{start} The matched position marks the START of one block
\item \code{end} The matched position marks the END of one block
}
}
}
}
\examples{
DT <- data.table(ID1 = c(NA, "A", "A", NA, NA, "B"), ID2 = c(rep("1", 4), "2", "2"), X1 = NA, X2 = c(NA, 1:2, NA, NA, 3), X3 = c(rep(NA, 5), .1))
tidy_hsplit(DT[, -2L])
tidy_hsplit(DT[, -2L], trim = TRUE)
tidy_hsplit(DT, cols = c("ID1", "ID2"))
}
\keyword{internal}
