% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{use_common_dir}
\alias{use_common_dir}
\title{Create common directory structure for analysis}
\usage{
use_common_dir(wd = NULL, flavor = c("op", "gc"), create_project = TRUE)
}
\description{
Create common directory structure for analysis
}
