% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_lookup.R
\name{vec_lookup}
\alias{vec_lookup}
\title{Excel \code{vlookup} equivalent}
\usage{
vec_lookup(
  x,
  key,
  value,
  method = c("exact", "regex"),
  ignore_case = TRUE,
  default = NULL
)
}
\description{
Excel \code{vlookup} equivalent
}
\examples{
(x <- c("A", "B", "C", "CBA", "abc"))
key <- c("a", "b", "A")
value <- c("Alice", "Bob", "Apple")
vec_lookup(x, key, value)
vec_lookup(x, key, value, ignore_case = FALSE)
vec_lookup(x, key, value, default = "Nobody")
vec_lookup(x, key, value, default = x)
vec_lookup(x, key, value, method = "regex")
vec_lookup(x, paste0("^", key, "$"), value, method = "regex")
}
