% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makenames.R
\name{vec_setnames}
\alias{vec_setnames}
\title{Quickly fixed vector names}
\usage{
vec_setnames(x, nm = x, overwrite = TRUE)
}
\description{
Quickly fixed vector names
}
\examples{
vec_setnames(1:3)
vec_setnames(c("a" = 1, 2))
vec_setnames(c("a" = 1, 2), overwrite = FALSE)
vec_setnames(list("a" = 1, 2), overwrite = FALSE)

}
