% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_format.R
\name{num}
\alias{num}
\alias{num_format}
\alias{num_auto}
\title{Format numbers}
\usage{
num(
  x,
  scale = c("C", "P", "K", "M", "B", "pp"),
  digits = 0,
  sep = "",
  na_as = NA_character_,
  inf_as = "Inf"
)

num_format(scale = c("C", "P", "K", "M", "B"), digits = 0, sep = "")

num_auto(
  x,
  prefer_digits = 0,
  free_digits = FALSE,
  max_digits = 4,
  perc_cutoff = 5,
  p = 0.95
)
}
\arguments{
\item{perc_cutoff}{The threshold for percentage format.  If most elements are lower than it, the whole vector will be formatted as percentage.}
}
\description{
Format numbers
}
\examples{
x <- c(1.23e12, 3.45e7, 6.78e3, 0.1, 0.012, 0.0123, pi / c(1, 100, 1000), Inf, NA)
num(x, digits = 4)
num(x, "P")
vapply(x, num_auto, character(1L))
vapply(x, num_auto, character(1L), perc_cutoff = 2)
num_auto(c(NA, 100, 1000, 10))
num_auto(c(NA, 100, 1000, 10), prefer_digits = 1)
num_auto(c(100e6, 200e6, 300e6))
num_auto(c(-123.456789e6, 1e9, 2e9, 0, -3e9))
num_auto(c(-123.456789e6, 1e9, 2e9, 0, -3e9), free_digits = TRUE, max_digits = 2)
}
