% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onepunch-package.R
\docType{package}
\name{onepunch-package}
\alias{onepunch}
\alias{onepunch-package}
\title{onepunch: An R toolkit for my everyday work}
\description{
`onepunch` is an R toolkit designed to accelerate modeling, pricing and risk-management analysis in reinsurance transactions
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/addnox/onepunch}
  \item Report bugs at \url{https://github.com/addnox/onepunch/issues}
}

}
\author{
\strong{Maintainer}: Charles Pan \email{addnox@gmail.com}

}
\keyword{internal}
