% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_extra.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Plot waterfall chart}
\usage{
plot_waterfall(
  DT,
  balance_start = NULL,
  calc_balance = NULL,
  label_final = "Final Balance",
  color = c("#9ae5de", "#efe8d1", "#acc8d4"),
  x_var = NULL,
  y_var = NULL,
  bar_width = 0.5,
  show_hline = FALSE,
  show_label = FALSE,
  labeller = NULL
)
}
\description{
Plot waterfall chart
}
\examples{
df <- data.table(x = c("LR", "LPC", "Comm", "PC"), y = -c(.7, -.1, .5, 0))
plot_waterfall(df, show_hline = T, show_label = T)
plot_waterfall(df, balance_start = c("Premium" = 1), show_hline = T, show_label = T, labeller = num_auto)
plot_waterfall(df, balance_start = c("Premium" = 1), calc_balance = c("Bal after LPC" = "LPC", "Treaty Balance" = "PC"), show_hline = T, color = c("green", "red", "grey"))
}
