% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_split.R
\name{tidy_hsplit}
\alias{tidy_hsplit}
\alias{tidy_vsplit}
\title{Split data into list of \code{data.table}s horizontally or vertically}
\usage{
tidy_hsplit(x, cols = NULL, trim = FALSE, na.rm = TRUE, keep_cols = TRUE)

tidy_vsplit(x, rows = NULL, trim = FALSE, na.rm = TRUE, keep_rows = TRUE)
}
\description{
Split data into list of \code{data.table}s horizontally or vertically
}
